Add tests for the relocation inside modules of static/const data.
The static/const data, when linked inside a module causes the relocation
code to be invoked. This code was broken in the case of certain structs
which relocated relative to the readonly data (rodata) section. We
weren't referencing the destination section properly. This should now be
handled correctly.

I believe we may be able to separate the workspace from the module data,
but still cannot handle the case that the module is multiply
instantiated. Maybe this can come later, if we work out a way that the
relocations can function.