Updated AIF/RMF linkage for R/W region to be page aligned; added docs.
The AIF/RMF linkage now ensures that the R/W region starts on a 4096
byte boundary (page boundary). This will allow that region to be marked
as read-write, whilst the preceding region can be read-only. This does
not happen in the current OS, but it should be possible in the future.

The `riscos64-mkreloc` code now includes a description of the symbols
used and the table that is generated. The `__RelocCode` function can
also be replaced if someone wants to write a better relocation routine,
or one which has more diagnostics.