Update zero-init code to check that we have enough memory.
Before we try to write over lots of memory, we now check whether we
have enough space in the application slot. If we don't we raise an
error, rather than trying to initialise it and then crashing when
we hit the unallocated memory (or possibly overwriting the app at
the top of the slot - if we ever actually do things that way).