Add implementation of the fscanf/vfscanf functions.
The implementation of the vfscanf function has been taken from
an Android source (which is itself OpenBSD source) and updated to
be able to be used if there is no buffer available for file I/O
(as we don't have any here at present).

The sscanf that I wrote is much less featureful, and will probably
need replacing in a similar way at some point.