Update version number to 0.21.
A couple of new maths, time and string functions which were missing.
`fscanf`, too, is now implemented, which makes for a slightly odd
environment as `sscanf` isn't implemented properly at all, but `fscanf`
is.

`rand` in particular is now implemented properly - as it would previously
return the same number on each call. And `fputs` no longer writes a
newline after every call (mistake introduced when `puts` was fixed).